requirements.txt
django
djangorestframework
python-dotenv
django-ckeditor
django-cors-headers
gunicorn
psycopg2-binary
pillow
